+++ controllers/cloud.redhat.com/makers/maker.go
@@ -89,7 +89,8 @@ func (mu *makerUpdater) Updater() utils.Updater {
 	return mu.update
 }
 
-// GetNamespacedName returns a new types.NamespacedName from the request based on pattern
+// GetNamespacedName returns a new types.NamespacedName from the request based
+// on pattern
 func GetNamespacedName(r *reconcile.Request, pattern, namespace string) types.NamespacedName {
 	if namespace == "" {
 		namespace = r.Namespace
