+++ controllers/cloud.redhat.com/makers/object_store.go
@@ -132,176 +132,3 @@ func (obs *ObjectStoreMaker) minio() error {
 	return nil
 }
 
-// MakeMinio creates the actual minio service to be used by clowdapps, this does not create buckets
-func MakeMinio(m *Maker) (ctrl.Result, error) {
-	result := ctrl.Result{}
-	nn := GetNamespacedName(m.Request, "%v-minio", m.Env.Spec.Namespace)
-
-	dd := apps.Deployment{}
-	update, err := m.Get(nn, &dd)
-	if err != nil {
-		return result, err
-	}
-
-	labels := m.Env.GetLabels()
-	labels["env-app"] = nn.Name
-
-	labeler := m.MakeLabeler(nn, labels)
-
-	labeler(&dd)
-
-	replicas := int32(1)
-
-	dd.Spec.Replicas = &replicas
-	dd.Spec.Selector = &metav1.LabelSelector{MatchLabels: labels}
-	dd.Spec.Template.Spec.Volumes = []core.Volume{{
-		Name: nn.Name,
-		VolumeSource: core.VolumeSource{
-			PersistentVolumeClaim: &core.PersistentVolumeClaimVolumeSource{
-				ClaimName: nn.Name,
-			},
-		}},
-	}
-	dd.Spec.Template.ObjectMeta.Labels = labels
-
-	dd.Spec.Template.Spec.ImagePullSecrets = []core.LocalObjectReference{{
-		Name: "quay-cloudservices-pull",
-	}}
-
-	secret := &core.Secret{}
-	secretUpdate, err := m.Get(nn, secret)
-	if err != nil {
-		return result, err
-	}
-
-	if len(secret.Data) == 0 {
-		hostname := fmt.Sprintf("%v.%v.svc", nn.Name, nn.Namespace)
-		port := int32(9000)
-		secret.StringData = map[string]string{
-			"accessKey": utils.RandString(12),
-			"secretKey": utils.RandString(12),
-			"hostname":  hostname,
-			"port":      strconv.Itoa(int(port)),
-		}
-
-		secret.Name = nn.Name
-		secret.Namespace = nn.Namespace
-		secret.ObjectMeta.OwnerReferences = []metav1.OwnerReference{m.Env.MakeOwnerReference()}
-		secret.Type = core.SecretTypeOpaque
-
-		if result, err = secretUpdate.Apply(secret); err != nil {
-			return result, err
-		}
-
-		m.Env.Status.ObjectStore = crd.ObjectStoreStatus{
-			Buckets: []string{},
-			Minio: crd.MinioStatus{
-				Credentials: core.SecretReference{
-					Name:      secret.Name,
-					Namespace: secret.Namespace,
-				},
-				Hostname: hostname,
-				Port:     port,
-			},
-		}
-
-		err = m.Client.Status().Update(m.Ctx, m.Env)
-
-		if err != nil {
-			return result, err
-		}
-	}
-
-	envVars := []core.EnvVar{{
-		Name: "MINIO_ACCESS_KEY",
-		ValueFrom: &core.EnvVarSource{
-			SecretKeyRef: &core.SecretKeySelector{
-				LocalObjectReference: core.LocalObjectReference{
-					Name: nn.Name,
-				},
-				Key: "accessKey",
-			},
-		},
-	}, {
-		Name: "MINIO_SECRET_KEY",
-		ValueFrom: &core.EnvVarSource{
-			SecretKeyRef: &core.SecretKeySelector{
-				LocalObjectReference: core.LocalObjectReference{
-					Name: nn.Name,
-				},
-				Key: "secretKey",
-			},
-		},
-	}}
-
-	ports := []core.ContainerPort{{
-		Name:          "minio",
-		ContainerPort: 9000,
-	}}
-
-	// TODO Readiness and Liveness probes
-
-	c := core.Container{
-		Name:  nn.Name,
-		Image: "minio/minio",
-		Env:   envVars,
-		Ports: ports,
-		VolumeMounts: []core.VolumeMount{{
-			Name:      nn.Name,
-			MountPath: "/storage",
-		}},
-		Args: []string{
-			"server",
-			"/storage",
-		},
-	}
-
-	dd.Spec.Template.Spec.Containers = []core.Container{c}
-	dd.Spec.Template.SetLabels(labels)
-
-	if result, err = update.Apply(&dd); err != nil {
-		return result, err
-	}
-
-	s := core.Service{}
-	update, err = m.Get(nn, &s)
-	if err != nil {
-		return result, err
-	}
-
-	servicePorts := []core.ServicePort{{
-		Name:     "minio",
-		Port:     9000,
-		Protocol: "TCP",
-	}}
-
-	labeler(&s)
-
-	s.Spec.Selector = labels
-	s.Spec.Ports = servicePorts
-
-	if result, err = update.Apply(&s); err != nil {
-		return result, err
-	}
-
-	pvc := core.PersistentVolumeClaim{}
-
-	update, err = m.Get(nn, &pvc)
-	if err != nil {
-		return result, err
-	}
-
-	labeler(&pvc)
-
-	pvc.Spec.AccessModes = []core.PersistentVolumeAccessMode{core.ReadWriteOnce}
-	pvc.Spec.Resources = core.ResourceRequirements{
-		Requests: core.ResourceList{
-			core.ResourceName(core.ResourceStorage): resource.MustParse("1Gi"),
-		},
-	}
-
-	if result, err = update.Apply(&pvc); err != nil {
-		return result, err
-	}
-	return result, nil
-}
