--- controllers/cloud.redhat.com/makers/objectstore/objectstore.go
@@ -1,7 +1,0 @@
-package objectstore
-
-import "context"
-
-type ObjectStore interface {
-	CreateBucket(ctx context.Context, bucket string) error
-}
